refactor(guideline): rewrite guideline as an optional behavioural attribute directive on vzDesigner
the controller of vzDesigner now offers a set of api points for setting up interceptors for element dragging. an interceptor receives the point to which the dragging element is moved, and possibly changes it to some new point. by this approach, other more complicated logic depending on the application can also be implemented instead of the implemented guideline, to help user in moving elements.